+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # pibackup - CHANGELOG
 
+- 0.5 - My favorite update so far
+  - Fix #9 (no need to run from remote node)
+  - Fix #4 (merge rotate into main script)
+  - Add a cheap-ass logo from "free" stuff
+  - Update doc accordingly to new features
 - 0.4.1 - Fix remote logic
   - Variables were not correctly set:used. Still subject to modification, but so far the correct logic is now implemented.
 - 0.4 - Compression bugfix